+++ custom_components/tuya_ble/switch.py
@@ -323,24 +323,6 @@ mapping: dict[str, TuyaBLECategorySwitchMapping] = {
                             entity_category=EntityCategory.CONFIG,
                         ),
                     ),
-                    #TODO:
-                    TuyaBLESwitchMapping( 
-                        dp_id=107,
-                        description=SwitchEntityDescription(
-                            key="programming_mode",
-                            icon="mdi:calendar-edit",
-                            entity_category=EntityCategory.CONFIG,
-                        ),
-                    ),
-                    #TODO:
-                    TuyaBLESwitchMapping(
-                        dp_id=108,
-                        description=SwitchEntityDescription(
-                            key="programming_switch",
-                            icon="mdi:calendar-clock",
-                            entity_category=EntityCategory.CONFIG,
-                        ),
-                    ),
                 ],
             ),
         },